refactor: app 모듈 컨트롤러 @MVC 컨트롤러로 변경 및 `asis` 패키지 레거시 코드 삭제
