+++ README.md
@@ -29,6 +29,6 @@
 - [x] Legacy MVC와 @MVC 통합하기
 - [ ] JsonView 클래스를 구현
 - [ ] Legacy MVC 제거
-  - [ ] app 모듈에 있는 모든 컨트롤러를 어노테이션 기반 MVC로 변경
-  - [ ] asis 패키지에 있는 레거시 코드를 삭제해도 서비스가 정상 동작하도록 리팩터링
+  - [x] app 모듈에 있는 모든 컨트롤러를 어노테이션 기반 MVC로 변경
+  - [x] asis 패키지에 있는 레거시 코드를 삭제해도 서비스가 정상 동작하도록 리팩터링
   - [ ] Legacy MVC를 제거하고 나서 DispatcherServlet도 app 패키지가 아닌 mvc 패키지로 이동
