+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -1,37 +1,52 @@
 package com.techcourse.controller;
 
+import com.interface21.context.stereotype.Controller;
+import com.interface21.web.bind.annotation.RequestMapping;
+import com.interface21.web.bind.annotation.RequestMethod;
+import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.servlet.view.JspView;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LoginController implements Controller {
+@Controller
+public class LoginController {
 
     private static final Logger log = LoggerFactory.getLogger(LoginController.class);
 
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        if (UserSession.isLoggedIn(req.getSession())) {
-            return "redirect:/index.jsp";
-        }
-
-        return InMemoryUserRepository.findByAccount(req.getParameter("account"))
+    @RequestMapping(value = "/login", method = RequestMethod.GET)
+    public ModelAndView show(HttpServletRequest request, HttpServletResponse response) {
+        return UserSession.getUserFrom(request.getSession())
                 .map(user -> {
-                    log.info("User : {}", user);
-                    return login(req, user);
+                    log.info("logged in {}", user.getAccount());
+                    return new ModelAndView(new JspView("redirect:/index.jsp"));
                 })
-                .orElse("redirect:/401.jsp");
+                .orElse(new ModelAndView(new JspView("/login.jsp")));
     }
 
-    private String login(final HttpServletRequest request, final User user) {
+    @RequestMapping(value = "/login", method = RequestMethod.POST)
+    public ModelAndView login(HttpServletRequest request, HttpServletResponse response) {
+        if (UserSession.isLoggedIn(request.getSession())) {
+            return new ModelAndView(new JspView("redirect:/index.jsp"));
+        }
+
+        return InMemoryUserRepository.findByAccount(request.getParameter("account"))
+                .map(user -> {
+                    log.info("User : {}", user);
+                    return login(request, user);
+                })
+                .orElse(new ModelAndView(new JspView("redirect:/401.jsp")));
+    }
+
+    private ModelAndView login(final HttpServletRequest request, final User user) {
         if (user.checkPassword(request.getParameter("password"))) {
             final var session = request.getSession();
             session.setAttribute(UserSession.SESSION_KEY, user);
-            return "redirect:/index.jsp";
+            return new ModelAndView(new JspView("redirect:/index.jsp"));
         }
-        return "redirect:/401.jsp";
+        return new ModelAndView(new JspView("redirect:/401.jsp"));
     }
 }
