+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerAdapterRegistryTest.java
@@ -4,12 +4,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.mockito.Mockito.mock;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.webmvc.servlet.mvc.AnnotationHandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.HandlerAdapterRegistry;
+import com.interface21.webmvc.servlet.mvc.HandlerExecution;
 import com.interface21.webmvc.servlet.view.JspView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import samples.TestManualHandlerAdapter;
 
 class HandlerAdapterRegistryTest {
 
@@ -19,7 +20,6 @@ class HandlerAdapterRegistryTest {
     void setUp() {
         registry = new HandlerAdapterRegistry();
         registry.addHandlerAdapter(new AnnotationHandlerAdapter());
-        registry.addHandlerAdapter(new TestManualHandlerAdapter());
     }
 
     @DisplayName("HandlerExecution 인스턴스를 전달하면 AnnotationHandlerAdapter를 반환한다.")
@@ -28,14 +28,6 @@ class HandlerAdapterRegistryTest {
         HandlerExecution handler = mock(HandlerExecution.class);
 
         assertThat(registry.getHandlerAdapter(handler)).isInstanceOf(AnnotationHandlerAdapter.class);
-    }
-
-    @DisplayName("Controller 인스턴스를 전달하면 ManualHandlerAdapter를 반환한다.")
-    @Test
-    void getManualHandlerAdapter() {
-        Controller handler = mock(Controller.class);
-
-        assertThat(registry.getHandlerAdapter(handler)).isInstanceOf(TestManualHandlerAdapter.class);
     }
 
     @DisplayName("적합한 핸들러 어댑터가 없으면 예외가 발생한다.")
