--- mvc/src/test/java/samples/TestManualHandlerAdapter.java
@@ -1,24 +1,0 @@
-package samples;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.View;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
-import com.interface21.webmvc.servlet.view.JspView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class TestManualHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        String viewName = ((Controller) handler).execute(request, response);
-        View jspView = new JspView(viewName);
-        return new ModelAndView(jspView);
-    }
-}
