--- mvc/src/test/java/samples/TestManualHandlerMapping.java
@@ -1,24 +1,0 @@
-package samples;
-
-import java.util.HashMap;
-import java.util.Map;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.asis.ForwardController;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
-import jakarta.servlet.http.HttpServletRequest;
-
-public class TestManualHandlerMapping implements HandlerMapping {
-
-    private static final Map<String, Controller> controllers = new HashMap<>();
-
-    public void initialize() {
-        controllers.put("/", new ForwardController("/index.jsp"));
-        controllers.put("/get-test", new TestLegacyController());
-    }
-
-    @Override
-    public Object getHandler(HttpServletRequest request) {
-        String requestURI = request.getRequestURI();
-        return controllers.get(requestURI);
-    }
-}
